Completely modularize code
