+++ src/components/append.js
@@ -3,7 +3,6 @@ import {showDetailsButtonListener} from './todoitem';
 import {removeButtonListener} from './todoitem';
 import {editButtonListener} from './todoitem';
 import {editFormListener} from './todoitem';
-import {bla} from './todoitem';
 
 const appendElements = (arr, container) => {
   for (let i = 0; i < arr.length; i++) {
@@ -13,7 +12,6 @@ const appendElements = (arr, container) => {
   removeButtonListener(arr, container);
   editButtonListener(container);
   editFormListener(arr, container);
-  bla();
 };
 
 export default appendElements;
